backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
aixplain>=1.0.0  # Add the aixplain SDK